refactor(dashboard): extract auth-import modals — #3501 Phase 1a
Strangler-fig decomposition of providers/[id], building on Phase 0. Extracts
the three self-contained auth-import modal clusters into components/modals/:

- ImportCodexAuthModal + ApplyCodexAuthModal (+ co-located helpers)
- ImportClaudeAuthModal + ApplyClaudeAuthModal
- ImportGeminiAuthModal + ApplyGeminiAuthModal

ProviderDetailPageClient.tsx: 12,882 -> 10,720 LOC. Each modal cluster had a
clean { onClose, onSuccess } / inline-prop interface and its own helpers, so the
move is structural only. Unused imports pruned (each new file < 800-LOC cap).

Validation: typecheck:core clean, lint 0 errors, prettier clean, the Phase 0
page smoke test still green (mount unaffected), and a new standalone render test
proves each extracted modal mounts on its own (Hard Rule #8). file-size ratchet
lowered for the client.

Part of #3501. Co-authored-by: oyi77 <oyi77@users.noreply.github.com>
